Run code cleanup

+++ core/src/main/java/org/nzbhydra/searching/SearchResultIdCalculator.java
@@ -65,10 +65,9 @@ public class SearchResultIdCalculator {
             case 2:
                 h ^= (long) (data[(length & ~7) + 1] & 0xff) << 8;
             case 1:
-                h ^= (long) (data[length & ~7] & 0xff);
+                h ^= data[length & ~7] & 0xff;
                 h *= m;
         }
-        ;
 
         h ^= h >>> r;
         h *= m;
